mysite/requirements.txt
asgiref==3.7.2
autopep8==2.3.1
dj-database-url==2.2.0
Django==5.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
